requirements.txt
cmd2==2.5.9
colorama==0.4.6
gnureadline==8.2.13
iniconfig==2.1.0
numpy==2.2.3
packaging==25.0
pluggy==1.6.0
PrettyPrintTree==2.0.1
pyperclip==1.9.0
pytest==8.3.5
wcwidth==0.2.13
